Override 'My account' page to disable editing of user details
User name, email and password are managed by Postgres comminity login.  It is
confusing when people try to edit them in Redmine.

Disable input form fields and hide 'Change password' button, but add an
explanatory banner.

The fix is incomplete, because controller still allows changing these details.
But to use this loophole, a user will need to perform some actions manually
and they were warned not to do that.